pyproject.toml
[project]
name = "tiny-spider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "adguardhome>=0.7.0",
    "aw-client>=0.5.15",
    "boto3>=1.35.65",
    "click>=8.1.8",
    "duckdb>=1.1.3",
    "fabric>=3.2.2",
    "infisicalsdk>=0.2.0",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pillow-heif>=0.21.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "s3fs>=2024.12.0",
    "schedule>=1.2.2",
    "structlog>=24.4.0",
]

[tool.uv.workspace]
members = ["ara"]
